+++ base/__init__.py
@@ -1,2 +1,2 @@
-VERSION = [1, 8, 16]
+VERSION = [1, 9, 0]
 __VERSION__ = '.'.join(map(str, VERSION))